requirements.txt
altgraph==0.17.4
asgiref==3.7.2
bleach==6.1.0
Django==4.2.4
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
gunicorn==21.2.0
pycparser==2.21
PyScreeze==0.1.29
PySocks==1.7.1
pytweening==1.0.7
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1

